backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
cryptography>=41.0.0
django-ratelimit>=4.1.0
openai>=1.0.0
ollama>=0.1.0
celery>=5.3.0
redis>=5.0.0